Fix return all search results
In the specification of Rubygems API, in the case of a search result
over a plurality of pages,
the amount of data of the first page may be less than or equal to the
maximum number (maximum of 30).
Therefore, it searched for the next page and corrected to end when the
result is 0.